requirements.txt
numpy==1.14.3
pandas==0.23.0
-e git+https://github.com/pydata/pandas-datareader.git@master#egg=MyProject
firebase-admin==2.12.0
requests==1.1.0
python-firebase==1.2